remove more dead code and restructure
