--- src/net/aegistudio/mpp/algo/StringGenerator.java
@@ -1,10 +1,0 @@
-
-package net.aegistudio.mpp.algo;
-
-import net.aegistudio.mpp.export.Asset;
-
-public interface StringGenerator
-extends Asset {
-    public int string(Paintable var1, int var2, int var3, float var4, String var5);
-}
-
