--- src/net/aegistudio/mpp/canvas/CanvasWrapper.java
@@ -1,11 +1,0 @@
-
-package net.aegistudio.mpp.canvas;
-
-import java.util.Set;
-
-public interface CanvasWrapper {
-    public void setWrapping(int var1, String var2);
-
-    public void showWrapped(Set<String> var1);
-}
-
