--- src/net/aegistudio/mpp/canvas/DecalCanvas.java
@@ -1,12 +1,0 @@
-
-package net.aegistudio.mpp.canvas;
-
-import net.aegistudio.mpp.MapPainting;
-
-public class DecalCanvas
-extends WrapCanvas {
-    public DecalCanvas(MapPainting painting) {
-        super(painting);
-    }
-}
-
