--- src/net/aegistudio/mpp/export/Context.java
@@ -1,12 +1,0 @@
-
-package net.aegistudio.mpp.export;
-
-import net.aegistudio.mpp.algo.Paintable;
-
-public interface Context
-extends Paintable {
-    public void clear();
-
-    public void repaint();
-}
-
