--- src/net/aegistudio/mpp/export/PluginCanvasFactory.java
@@ -1,7 +1,0 @@
-
-package net.aegistudio.mpp.export;
-
-public interface PluginCanvasFactory<T extends PluginCanvas> {
-    public T create(Context var1);
-}
-
