+++ src/net/aegistudio/mpp/tool/PaintTool.java
@@ -1,6 +1,8 @@
 
-package net.aegistudio.mpp;
+package net.aegistudio.mpp.tool;
 
+import net.aegistudio.mpp.Interaction;
+import net.aegistudio.mpp.Module;
 import net.aegistudio.mpp.canvas.MapCanvasRegistry;
 import org.bukkit.inventory.ItemStack;
