+++ src/net/aegistudio/mpp/tool/PaintToolManager.java
@@ -1,9 +1,13 @@
 
-package net.aegistudio.mpp;
+package net.aegistudio.mpp.tool;
 
 import java.util.Map;
 import java.util.TreeMap;
+
+import net.aegistudio.mpp.Interaction;
+import net.aegistudio.mpp.MapPainting;
 import net.aegistudio.mpp.canvas.MapCanvasRegistry;
+
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.inventory.ItemStack;
